refactor: 🚚 Old SponsorCard moved inside `/components/Main`
Related: #175
