+++ src/components/Main/Sponsors/SponsorCard/index.tsx
